services/CTR1PANEL/requirements.txt
Flask==3.0.0
regex==2023.12.25
psycopg2-binary==2.9.9
flask-login==0.6.3
PyJWT==1.5.3

